Fixed further compiler issues with Marmalade
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@1493 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/add_on/scripthelper/scripthelper.cpp
@@ -191,7 +191,7 @@ int WriteConfigToFile(asIScriptEngine *engine, const char *filename)
 		const char *enumName = engine->GetEnumByIndex(n, &typeId, &nameSpace, 0, &accessMask);
 		if( accessMask != currAccessMask )
 		{
-			fprintf(f, "access %X\n", accessMask);
+			fprintf(f, "access %X\n", (unsigned int)(accessMask));
 			currAccessMask = accessMask;
 		}
 		if( nameSpace != currNamespace )
@@ -219,7 +219,7 @@ int WriteConfigToFile(asIScriptEngine *engine, const char *filename)
 		asDWORD accessMask = type->GetAccessMask();
 		if( accessMask != currAccessMask )
 		{
-			fprintf(f, "access %X\n", accessMask);
+			fprintf(f, "access %X\n", (unsigned int)(accessMask));
 			currAccessMask = accessMask;
 		}
 		const char *nameSpace = type->GetNamespace();
@@ -258,7 +258,7 @@ int WriteConfigToFile(asIScriptEngine *engine, const char *filename)
 		}
 		if( accessMask != currAccessMask )
 		{
-			fprintf(f, "access %X\n", accessMask);
+			fprintf(f, "access %X\n", (unsigned int)(accessMask));
 			currAccessMask = accessMask;
 		}
 		fprintf(f, "typedef %s \"%s\"\n", typeDef, engine->GetTypeDeclaration(typeId));
@@ -277,7 +277,7 @@ int WriteConfigToFile(asIScriptEngine *engine, const char *filename)
 		}
 		if( accessMask != currAccessMask )
 		{
-			fprintf(f, "access %X\n", accessMask);
+			fprintf(f, "access %X\n", (unsigned int)(accessMask));
 			currAccessMask = accessMask;
 		}
 		fprintf(f, "funcdef \"%s\"\n", funcDef->GetDeclaration());
@@ -305,7 +305,7 @@ int WriteConfigToFile(asIScriptEngine *engine, const char *filename)
 				asDWORD accessMask = func->GetAccessMask();
 				if( accessMask != currAccessMask )
 				{
-					fprintf(f, "access %X\n", accessMask);
+					fprintf(f, "access %X\n", (unsigned int)(accessMask));
 					currAccessMask = accessMask;
 				}
 				fprintf(f, "intfmthd %s \"%s\"\n", typeDecl.c_str(), func->GetDeclaration(false));
@@ -320,7 +320,7 @@ int WriteConfigToFile(asIScriptEngine *engine, const char *filename)
 				asDWORD accessMask = func->GetAccessMask();
 				if( accessMask != currAccessMask )
 				{
-					fprintf(f, "access %X\n", accessMask);
+					fprintf(f, "access %X\n", (unsigned int)(accessMask));
 					currAccessMask = accessMask;
 				}
 				fprintf(f, "objbeh \"%s\" %d \"%s\"\n", typeDecl.c_str(), asBEHAVE_FACTORY, func->GetDeclaration(false));
@@ -337,7 +337,7 @@ int WriteConfigToFile(asIScriptEngine *engine, const char *filename)
 				asDWORD accessMask = func->GetAccessMask();
 				if( accessMask != currAccessMask )
 				{
-					fprintf(f, "access %X\n", accessMask);
+					fprintf(f, "access %X\n", (unsigned int)(accessMask));
 					currAccessMask = accessMask;
 				}
 				fprintf(f, "objmthd \"%s\" \"%s\"\n", typeDecl.c_str(), func->GetDeclaration(false));
@@ -348,7 +348,7 @@ int WriteConfigToFile(asIScriptEngine *engine, const char *filename)
 				type->GetProperty(m, 0, 0, 0, 0, 0, &accessMask);
 				if( accessMask != currAccessMask )
 				{
-					fprintf(f, "access %X\n", accessMask);
+					fprintf(f, "access %X\n", (unsigned int)(accessMask));
 					currAccessMask = accessMask;
 				}
 				fprintf(f, "objprop \"%s\" \"%s\"\n", typeDecl.c_str(), type->GetPropertyDeclaration(m));
@@ -372,7 +372,7 @@ int WriteConfigToFile(asIScriptEngine *engine, const char *filename)
 		asDWORD accessMask = func->GetAccessMask();
 		if( accessMask != currAccessMask )
 		{
-			fprintf(f, "access %X\n", accessMask);
+			fprintf(f, "access %X\n", (unsigned int)(accessMask));
 			currAccessMask = accessMask;
 		}
 		fprintf(f, "func \"%s\"\n", func->GetDeclaration());
@@ -392,7 +392,7 @@ int WriteConfigToFile(asIScriptEngine *engine, const char *filename)
 		engine->GetGlobalPropertyByIndex(n, &name, &nameSpace, &typeId, &isConst, 0, 0, &accessMask);
 		if( accessMask != currAccessMask )
 		{
-			fprintf(f, "access %X\n", accessMask);
+			fprintf(f, "access %X\n", (unsigned int)(accessMask));
 			currAccessMask = accessMask;
 		}
 		if( nameSpace != currNamespace )
